+++ koru-processor/src/jvmTest/kotlin/com/futuremind/koru/processor/TypesGenerationTest.kt
@@ -1,10 +1,11 @@
 package com.futuremind.koru.processor
 
-import com.squareup.kotlinpoet.KModifier
 import com.tschuchort.compiletesting.KotlinCompilation
 import com.tschuchort.compiletesting.SourceFile
 import io.kotest.matchers.collections.shouldContain
 import io.kotest.matchers.shouldBe
+import io.kotest.matchers.string.shouldContain
+import io.kotest.matchers.string.shouldNotContain
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.io.TempDir
 import java.io.File
@@ -499,5 +500,73 @@ class TypesGenerationTest {
         expectedMessage = "Cannot wrap types with `private` modifier. Consider using internal or public.",
         tempDir = tempDir
     )
+
+    @Test
+    fun `should freeze wrapper if freeze=true in annotation`() {
+
+        val classToWrap = SourceFile.kotlin(
+            "freeze1.kt",
+            """
+                        package com.futuremind.kmm101.test
+                        
+                        import com.futuremind.koru.ToNativeClass
+                        import kotlinx.coroutines.flow.Flow
+                        
+                            @ToNativeClass(freeze = true)
+                            class FreezeExample {
+                                suspend fun suspending(whatever: Int) : Float = TODO()
+                                fun flow(whatever: Int) : Flow<Float> = TODO()
+                            }
+                    """
+        )
+
+        val compilationResult = prepareCompilation(
+            sourceFiles = listOf(classToWrap),
+            tempDir = tempDir
+        ).compile()
+
+        compilationResult.exitCode shouldBe KotlinCompilation.ExitCode.OK
+
+        val generatedClass = compilationResult.generatedFiles
+            .getContentByFilename("FreezeExample$defaultClassNameSuffix.kt")
+
+        generatedClass shouldContain "FlowWrapper<Float> = FlowWrapper(scopeProvider, true"
+        generatedClass shouldContain "SuspendWrapper(scopeProvider, true"
+        generatedClass shouldContain "this.freeze()"
+    }
+
+    @Test
+    fun `should not freeze wrapper by default`() {
+
+        val classToWrap = SourceFile.kotlin(
+            "freeze2.kt",
+            """
+                        package com.futuremind.kmm101.test
+                        
+                        import com.futuremind.koru.ToNativeClass
+                        import kotlinx.coroutines.flow.Flow
+                        
+                            @ToNativeClass
+                            class FreezeExample {
+                                suspend fun suspending(whatever: Int) : Float = TODO()
+                                fun flow(whatever: Int) : Flow<Float> = TODO()
+                            }
+                    """
+        )
+
+        val compilationResult = prepareCompilation(
+            sourceFiles = listOf(classToWrap),
+            tempDir = tempDir
+        ).compile()
+
+        compilationResult.exitCode shouldBe KotlinCompilation.ExitCode.OK
+
+        val generatedClass = compilationResult.generatedFiles
+            .getContentByFilename("FreezeExample$defaultClassNameSuffix.kt")
+
+        generatedClass shouldContain "FlowWrapper<Float> = FlowWrapper(scopeProvider, false"
+        generatedClass shouldContain "SuspendWrapper(scopeProvider, false"
+        generatedClass shouldNotContain "this.freeze()"
+    }
 
 }